Standard ID Provider - Remove use of services (#426)
