+++ modules/app-standardidprovider/src/main/resources/assets/js/login.js
@@ -21,8 +21,7 @@ function loginButtonClick() {
 
     var data = {
         user: $("#username-input").val(),
-        password: $("#password-input").val(),
-        userStore: CONFIG.userStoreKey
+        password: $("#password-input").val()
     };
     $.ajax({
         url: CONFIG.appLoginServiceUrl,
