+++ modules/app-standardidprovider/src/main/resources/idprovider/idprovider-config.txt
@@ -1,7 +1,6 @@
 <script>
     var CONFIG = {
         {{#redirectUrl}}redirectUrl: "{{{.}}}",{{/redirectUrl}}
-        userStoreKey: "{{userStoreKey}}",
         appLoginServiceUrl: "{{appLoginServiceUrl}}",
         messages: {{{messages}}}
     };
